requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic-settings==2.4.0
asyncpg==0.29.0
python-dotenv==1.0.1
pypdf==5.2.0
python-docx==1.1.2
httpx==0.27.2
msal==1.31.0
msgraph-sdk==1.11.0